commit rentacar1 modificare 15.08
