+++ src/main/java/ro/sci/rentacar1/IO/InvalidEntityException.java
@@ -4,6 +4,7 @@ package ro.sci.rentacar1.IO;
  * Created by Roxana on 8/7/2017.
  */
 public class InvalidEntityException extends Exception{
+
         public InvalidEntityException(String s) {
             super(s);
         }
