+++ src/main/java/ro/sci/rentacar1/repository/CarDataBase.java
@@ -1,0 +1,30 @@
+package ro.sci.rentacar1.repository;
+
+import ro.sci.rentacar1.domain.car.Car;
+
+import java.util.List;
+
+/**
+ * Created by Roxana on 8/15/2017.
+ */
+public class CarDataBase implements RepoInterface<Car> {
+    @Override
+    public void add(Car car) {
+
+    }
+
+    @Override
+    public List<Car> find() {
+        return null;
+    }
+
+    @Override
+    public void update(Car car) {
+
+    }
+
+    @Override
+    public void delete(Car car) {
+
+    }
+}
