--- src/main/java/ro/sci/rentacar1/repository/TransactionsRepo.java
@@ -1,30 +1,0 @@
-package ro.sci.rentacar1.repository;
-
-import ro.sci.rentacar1.domain.Transaction;
-
-import java.util.List;
-
-/**
- * Created by Roxana on 8/14/2017.
- */
-public class TransactionsRepo implements RepoInterface <Transaction> {
-    @Override
-    public void add(Transaction transaction) {
-
-    }
-
-    @Override
-    public List<Transaction> find() {
-        return null;
-    }
-
-    @Override
-    public void update(Transaction transaction) {
-
-    }
-
-    @Override
-    public void delete(Transaction transaction) {
-
-    }
-}
