+++ src/main/java/ro/sci/rentacar1/services/TransactionServInterface.java
@@ -1,0 +1,7 @@
+package ro.sci.rentacar1.services;
+
+/**
+ * Created by Roxana on 8/15/2017.
+ */
+public interface TransactionServInterface {
+}
